requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
Flask==3.0.3
gunicorn==22.0.0
google-api-python-client==2.169.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
PyNaCl==1.5.0
